LayoutTests/media/video-size-expected.txt

EXPECTED (video.videoWidth == '0') OK
EXPECTED (video.videoHeight == '0') OK
EVENT(loadedmetadata)
EXPECTED (video.videoWidth > '0') OK
EXPECTED (video.videoHeight > '0') OK
END OF TEST
